requirements.txt
# core
fastapi
uvicorn

# database
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings

# environments
PyYAML